Adjust the crates docs to match the new pitch
Unlike Bleach, the library I originally had in mind, Ammonia does not sanitize "user input with some HTML."
It sanitizes HTML. Ammonia is intended to be compliant with the HTML5 standard for parsing and serializing.

If you want to handle "user input with some HTML," run it through a pre-processor first.