Adding `action`, `formaction` and `ping` to `is_url_attr`
The `is_url_attr` function is used to decide whether elements need their
URLs filtered to comply users that want all links to remain on-origin.
This patch adds missing the attributes (`action` for `<form>`,
`formaction` for `<input>` and `<button>` and `ping` for `<a>`).

I have used the (non-normative) reference in the HTML spec to ensure
comprehensivenes:
https://html.spec.whatwg.org/multipage/indices.html#attributes-3