Make sure that relative URLs actually work when filtering
It's possible for `Url::parse("xxx") == Err(url::ParseError::RelativeUrlWithoutBase)` to evaluate to true,
but for the eventual `url.join("xxx")` to still fail with a parse erro, depending on what "xxx" actually is.
To detect this, we check that joining succeeds in addition to the normal checks for relative URLs.

This PR also runs "cargo fmt".

Fixes #136